// web/src/api/generated/BlockLite.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Amount } from "./Amount";
import type { BlockKind } from "./BlockKind";
import type { Hash32 } from "./Hash32";
import type { NodeId } from "./NodeId";
import type { PayoutDto } from "./PayoutDto";

/**
 * Block list row.
 */
export type BlockLite = { height: number, hash: Hash32, time_ms: number, size: number, tx_count: number, kind: BlockKind, producer: NodeId | null, payouts: Array<PayoutDto>, reward: Amount, fees: Amount, confirm_count: number, start_count: number, transfer_count: number, };